Change lower bound of DAC samples frequency. Fix some warnings.
git-svn-id: https://zxtune.googlecode.com/svn/trunk@2628 e19a5227-2556-0410-816f-a35683127991

+++ src/sound/impl/gainer.cpp
@@ -81,7 +81,7 @@ namespace Sound
       }
     }
   private:
-    typedef Math::FixedPoint<int64_t, 1u << 31> Coeff;
+    typedef Math::FixedPoint<int64_t, int64_t(1) << 31> Coeff;
     BOOST_STATIC_ASSERT(8 * sizeof(Coeff) >= boost::static_log2<Coeff::PRECISION>::value + Sample::BITS);
     Coeff Level;
     Coeff Step;
